restructured modules and code
