+++ app/src/main/java/com/likeminds/feedsx/branding/model/Font.kt
@@ -1,4 +1,4 @@
-package com.likeminds.feedsx
+package com.likeminds.feedsx.branding.model
 
 import androidx.annotation.Keep
